+++ packages/icestark-app/src/AppLink.tsx
@@ -16,14 +16,17 @@ const AppLink = (props: AppLinkProps) => {
     <a
       {...rest}
       href={linkTo}
-      onClick={e => {
+      onClick={(e) => {
         e.preventDefault();
         // eslint-disable-next-line no-alert
         if (message && window.confirm(message) === false) {
           return false;
         }
 
-        const changeState = window.history[replace ? 'replaceState' : 'pushState'];
+        /*
+        * Bind `replaceState` and `pushState` to window to avoid illegal invocation error
+         */
+        const changeState = window.history[replace ? 'replaceState' : 'pushState'].bind(window);
 
         changeState({}, null, linkTo);
       }}