+++ tests/index.spec.tsx
@@ -256,10 +256,10 @@ describe('AppRouter', () => {
 
     fireEvent.click(getByText(/Jump Hash/i));
     // url do not change, will not trigger onRouteChange
-    expect(props.onRouteChange).toHaveBeenCalledTimes(2);
+    expect(props.onRouteChange).toHaveBeenCalledTimes(1);
 
     fireEvent.click(getByText(/Jump 404/i));
-    expect(props.onRouteChange).toHaveBeenCalledTimes(3);
+    expect(props.onRouteChange).toHaveBeenCalledTimes(2);
     expect(props.onAppLeave).toHaveBeenCalledTimes(2);
     expect(container.innerHTML).toContain('NotFound');